+++ lib/entities/reorderable_entity.dart
@@ -54,6 +54,22 @@ class ReorderableEntity {
         hasSwappedOrder: false,
       );
 
+  factory ReorderableEntity.orientationChanged({
+    required ValueKey key,
+    required int updatedOrderId,
+    required int originalOrderId,
+  }) =>
+      ReorderableEntity(
+        key: key,
+        originalOrderId: originalOrderId,
+        updatedOrderId: updatedOrderId,
+        originalOffset: Offset.zero,
+        updatedOffset: Offset.zero,
+        size: Size.zero,
+        isBuildingOffset: true,
+        hasSwappedOrder: false,
+      );
+
   @override
   bool operator ==(Object other) {
     return other is ReorderableEntity &&